fixup! refactor filters, redirects and caching
