+++ lib/handler.js
@@ -3,8 +3,9 @@ import { json } from 'http-responders'
 import Sentry from '@sentry/node'
 import { URLSearchParams } from 'node:url'
 
+import { fetchRetrievalSuccessRate } from './stats-fetchers.js'
+
 /**
- *
  * @param {object} args
  * @param {import('pg').Pool} args.pgPool
  * @param {import('./typings').Logger} args.logger
@@ -34,7 +35,12 @@ const handler = async (req, res, pgPool) => {
   const { pathname, searchParams } = new URL(req.url, 'http://127.0.0.1')
   const segs = pathname.split('/').filter(Boolean)
   if (req.method === 'GET' && segs[0] === 'retrieval-success-rate' && segs.length === 1) {
-    await getRetrievalSuccessRate(pathname, searchParams, res, pgPool)
+    await getStatsWithFilterAndCaching(
+      pathname,
+      searchParams,
+      res,
+      pgPool,
+      fetchRetrievalSuccessRate)
   } else {
     notFound(res)
   }
@@ -63,31 +69,6 @@ const notFound = (res) => {
   res.end('Not Found')
 }
 
-/**
- * @param {string} pathname
- * @param {URLSearchParams} searchParams
- * @param {import('node:http').ServerResponse} res
- * @param {import('pg').Pool} pgPool
- */
-const getRetrievalSuccessRate = async (pathname, searchParams, res, pgPool) => {
-  const filter = parseAndValidateFilter(pathname, searchParams, res)
-  if (res.headersSent) return
-
-  setCacheControlForStatsResponse(res, filter)
-
-  // Fetch the "day" (DATE) as a string (TEXT) to prevent node-postgres for converting it into
-  // a JavaScript Date with a timezone, as that could change the date one day forward or back.
-  const { rows } = await pgPool.query(
-    'SELECT day::text, total, successful FROM retrieval_stats WHERE day >= $1 AND day <= $2',
-    [filter.from, filter.to]
-  )
-  const stats = rows.map(r => ({
-    day: r.day,
-    success_rate: r.total > 0 ? r.successful / r.total : null
-  }))
-  json(res, stats)
-}
-
 const getDayAsISOString = (d) => d.toISOString().split('T')[0]
 
 export const today = () => getDayAsISOString(new Date())
@@ -96,9 +77,10 @@ export const today = () => getDayAsISOString(new Date())
  * @param {string} pathname
  * @param {URLSearchParams} searchParams
  * @param {import('node:http').ServerResponse} res
- * @returns {{from: string | undefined; to: string | undefined}}
+ * @param {import('pg').Pool} pgPool
+ * @param {(import('pg').Pool, import('./typings').Filter) => Promise<object[]>} fetchStatsFn
  */
-export const parseAndValidateFilter = (pathname, searchParams, res) => {
+const getStatsWithFilterAndCaching = async (pathname, searchParams, res, pgPool, fetchStatsFn) => {
   let from = searchParams.get('from')
   let to = searchParams.get('to')
   let shouldRedirect = false
@@ -145,13 +127,16 @@ export const parseAndValidateFilter = (pathname, searchParams, res) => {
     return { from, to }
   }
 
-  // We have well-formed from & to dates now
-  return { from, to }
+  // We have well-formed from & to dates now, let's fetch the requested stats from the DB
+  const filter = { from, to }
+  const stats = await fetchStatsFn(pgPool, filter)
+  setCacheControlForStatsResponse(res, filter)
+  json(res, stats)
 }
 
 /**
  * @param {import('node:http').ServerResponse} res
- * @param {{from: string, to: string}} filter
+ * @param {import('./typings').Filter} filter
  */
 const setCacheControlForStatsResponse = (res, filter) => {
   // We cannot simply compare filter.to vs today() because there may be a delay in finalizing
