+++ lib/typings.d.ts
@@ -3,3 +3,8 @@ export interface Logger {
   error: typeof console.error;
   request: typeof console.info;
 }
+
+export interface Filter {
+  from: string;
+  to: string;
+}
